feat: Add a server package
This commit includes a thin wrapper around the http.Server type that
comes with a constructor leaning heavily on sane defaults. The provided
implementation handles graceful shutdown as a result of calling Done()
on the provided context. Finally, this implementation returns a nil
error if the server shuts down gracefully, only returning an error when
something goes wrong.