feat: Add a hash package
This commit includes a thin wrapper around the functions in the
golang.org/x/crypto library that make it a little easier to work with if
you're expecting to be passing around strings. It also hard-codes using
the default cost from the bcrypt library to remove the need to specify
it every time.