feat: Add a new Hash type
This commit modifies the hash package to work with a new custom type.
Under the hood, the Hash type just stores the same byte slice that
bcrypt.GenerateFromPassword returns, but it also allows us to hang
methods on the value. Most significantly, the Hash type implements the
slog.Valuer interface and returns a static string to prevent hashes from
accidentally leaking into logs.